+++ version_service.go
@@ -1,6 +1,6 @@
 package main
 
-const AppVersion = "v2.8.12"
+const AppVersion = "v2.8.13"
 
 type VersionService struct {
 	version string